Fixed major security flaw
Fixed a security bug that allowed a user to gain control of another id from the settings page if they knew the email of any other user.

I switched from the email to the mongodb ID as an identifier to the server to retrieve the current user.

This is more secure since mongodb generates ids using the timestamp of generation and 5 byte random value making it virtually impossible to guess.