Settings.js (#45)
* Added Settings Page

Everything seems to be working fine! All the fields on the settings page can be changed

* Lint fix

* Updated settings form layout

* Update form layout and some functionality changes

Password is not required anymore to update personal info fields

* Update settings.js

* Added a few comments

* Multiple messages added and password invalidation doesn't prevent other fields from saving.

* Minor bug fixed

* Update settings.js

* Firstname and lastname can now be empty

* Fixed major security flaw

Fixed a security bug that allowed a user to gain control of another id from the settings page if they knew the email of any other user.

I switched from the email to the mongodb ID as an identifier to the server to retrieve the current user.

This is more secure since mongodb generates ids using the timestamp of generation and 5 byte random value making it virtually impossible to guess.

* Removed frontend involvement in user identification altogether

The settings page is now as secure as can be. The server itself handles identifying the user now

* Update settings.js